feat: remove copied over files from `wm`
