+++ Cargo.toml
@@ -5,8 +5,13 @@ default-members = ["packages/wm", "packages/watcher"]
 
 [workspace.dependencies]
 anyhow = { version = "1", features = ["backtrace"] }
+clap = { version = "4", features = ["derive"] }
+futures-util = "0.3"
+serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1", features = ["raw_value"] }
 tokio = { version = "1", features = ["full"] }
 tauri-winres = "0.1"
+tokio-tungstenite = "0.21"
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+uuid = { version = "1", features = ["v4", "serde"] }
