+++ packages/wm-common/src/lib.rs
@@ -4,7 +4,9 @@ mod color;
 mod direction;
 mod display_state;
 mod dtos;
+mod ipc;
 mod length_value;
+mod opacity_value;
 mod point;
 mod rect;
 mod rect_delta;
@@ -20,7 +22,9 @@ pub use color::*;
 pub use direction::*;
 pub use display_state::*;
 pub use dtos::*;
+pub use ipc::*;
 pub use length_value::*;
+pub use opacity_value::*;
 pub use point::*;
 pub use rect::*;
 pub use rect_delta::*;
