+++ packages/wm-ipc-client/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "wm-ipc-client"
+version = "0.0.0"
+edition = "2021"
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+wm-common = { path = "../wm-common" }
+anyhow = { workspace = true }
+tokio = { workspace = true }
+tokio-tungstenite = { workspace = true }
+futures-util = { workspace = true }
+serde_json = { workspace = true }
+uuid = { workspace = true }
