--- packages/wm/src/common/point.rs
@@ -1,6 +1,0 @@
-/// Represents an x-y coordinate.
-#[derive(Debug, Clone)]
-pub struct Point {
-  pub x: i32,
-  pub y: i32,
-}
